// app/presentaciones3/src/interfaces/FormasEnvioViewModel.ts
export interface FormasEnvioViewModel {
  foen_Id: number;
  foen_Codigo: string;
  foen_Descripcion: string;
  foen_FechaCreacion: string;
  foen_FechaModificacion: string;
  usua_UsuarioCreacion: number;
  usua_UsuarioModificacion: number;
  usua_UsuarioEliminacion:number;
  foen_FechaEliminacion: string;
}